Added rotating a 3x3 matrix solution and made a new package to store future arrays related problems, also moved some other array solutions into said package
