+++ challenges/test/lib/src/test/java/code/LibraryTest.java
@@ -3,10 +3,10 @@
  */
 package code;
 
+import code.arrays.ArrayShift;
 import org.junit.Test;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 import static org.junit.Assert.*;
