+++ challenges/test/lib/src/test/java/code/arraystest/ArrayShiftTest.java
@@ -1,5 +1,6 @@
-package code;
+package code.arraystest;
 
+import code.arrays.ArrayShift;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
